Added class ProjectService
