+++ src/main/java/com/Portafolio/apiController/ProjectController.java
@@ -1,6 +1,8 @@
 package com.Portafolio.apiController;
 
 import com.Portafolio.domain.model.Project;
+import com.Portafolio.domain.service.ProjectService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -10,17 +12,15 @@ import java.util.List;
 @RestController
 @RequestMapping(path = "api/v1/projects")
 public class ProjectController {
+    private final ProjectService projectService;
+    @Autowired
+
+    public ProjectController(ProjectService projectService) {
+        this.projectService = projectService;
+    }
+
     @GetMapping
-    public List<Project> getJob(){
-        return List.of(
-                new Project(
-                        1,
-                        "Mi project 1",
-                        "Description",
-                        null,
-                        "http://ejemplo.com/1",
-                        "http://ejemplo.com/2"
-                )
-        );
+    public List<Project> getJobController(){
+        return projectService.getJob();
     }
 }
